# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for VK Intern test task",
        "title": "REST API for VK",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/auth/signIn": {
            "post": {
                "description": "\"Authentificates admin and returns JWT-token\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "\"Authentificate admin\"",
                "parameters": [
                    {
                        "description": "Adminname and password",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Admin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT-token",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/signUp": {
            "post": {
                "description": "\"Registers new admin\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "\"Registration for admins\"",
                "parameters": [
                    {
                        "description": "Adminname and password",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Admin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Admin": {
            "type": "object",
            "required": [
                "adminname",
                "password"
            ],
            "properties": {
                "adminname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}